requirements.txt
ansicon==1.89.0
arrow==1.3.0
asgiref==3.9.1
blessed==1.21.0
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
django-picklefield==3.3
django-q==1.3.9
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
filelock==3.18.0
fsspec==2025.7.0
Jinja2==3.1.6
jinxed==1.3.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python-headless==4.12.0.88
pillow==11.3.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==3.5.3
celery==5.3.6  # Versión estable (Python 3.8+)
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
types-python-dateutil==2.9.0.20250708
typing_extensions==4.14.1
tzdata==2025.2
wcwidth==0.2.13
ultralytics>=8.0.0
pandas>=2.0.0
shapely>=2.0.0
tqdm>=4.66.0

